+++ pmm-app/package.json
@@ -40,6 +40,8 @@
     "@grafana/data": "7.1.3",
     "@grafana/runtime": "7.1.3",
     "@grafana/ui": "7.1.3",
+    "@emotion/core": "^10.0.35",
+    "@types/jest": "^26.0.10",
     "antd": "^3.24.3",
     "axios": "^0.19.0",
     "d3": "^5.16.0",
