+++ pmm-app/src/pmm-settings/Settings.types.ts
@@ -22,5 +22,5 @@ export enum TabKeys {
   advanced = 'advanced-settings',
   ssh = 'ssh-key',
   alertManager = 'am-integration',
-  signUp = 'signUp',
+  perconaPlatform = 'perconaPlatform',
 }
