+++ pmm-app/src/pmm-settings/components/PlatformLogin/FieldAdapters/Field.tsx
@@ -1,11 +1,7 @@
 import React from 'react';
-import {
-  css, cx
-} from 'emotion';
+import { css, cx } from 'emotion';
 import { GrafanaTheme } from '@grafana/data';
-import {
-  stylesFactory, useTheme
-} from '@grafana/ui';
+import { stylesFactory, useTheme } from '@grafana/ui';
 
 
 export interface FieldProps {
