+++ pmm-app/src/pmm-settings/components/PlatformLogin/PlatformLogin.messages.ts
@@ -9,8 +9,10 @@ export const Messages = {
   signUpSucceeded: 'You have successfully created your credentials',
   loginSucceeded: 'You are logged in as',
   passwordLabel: 'Password',
+  confirmPasswordLabel: 'Confirm Password',
   emailLabel: 'Email',
   errors: {
     signUpFailed: 'There was a problem during the sign up process, please try again',
+    passwordMismatch: 'Passwords should match'
   },
 };
