+++ pmm-app/src/pmm-settings/components/PlatformLogin/PlatformLogin.test.tsx
@@ -6,18 +6,18 @@ import { LoggedIn } from './LoggedIn/LoggedIn';
 jest.mock('shared/components/helpers/notification-manager');
 
 describe('SignUp::', () => {
-  it('Should show a sign up form if an undefined email is passed', () => {
+  it('Should show a login form if an undefined email is passed', () => {
     const root = mount(<PlatformLogin userEmail={undefined} />);
     const loggedInEmail = root.find(LoggedIn).at(0);
-    const submitFrom = root.find('[data-qa="sign-up-form"]').at(0);
+    const submitFrom = root.find('[data-qa="sign-in-form"]').at(0);
 
     expect(loggedInEmail).toHaveLength(0);
     expect(submitFrom).toHaveLength(1);
   });
 
-  it('Should have the "Sign up" button disabled when first rendered', () => {
+  it('Should have the "Sign up" button when first rendered', () => {
     const root = mount(<PlatformLogin userEmail={undefined} />);
-    const submitButton = root.find('[data-qa="sign-up-submit-button"]').at(0);
+    const submitButton = root.find('[data-qa="sign-in-to-sign-up-button"]').at(0);
 
     expect(submitButton.props()).toHaveProperty('disabled');
   });
